improvements for xapi del cronjob

+++ Modules/CmiXapi/classes/class.ilCmiXapiDelModel.php
@@ -88,12 +88,13 @@ class ilCmiXapiDelModel
         ]);
     }
 
-    public function resetUpdatedXapiUser(int $usrId)
+    public function resetUpdatedXapiUser(int $usrId, int $objId)
     {
         $this->db->update(self::DB_DEL_USERS, [
             'updated' => ['timestamp', null]
         ], [
-            'usr_id' => ['integer', $usrId]
+            'usr_id' => ['integer', $usrId],
+            'obj_id' => ['integer', $objId]
         ]);
     }
 
@@ -145,12 +146,12 @@ class ilCmiXapiDelModel
         return $data;
     }
 
-    public function deleteUserEntry($usrId)
+    public function deleteUserEntry($usrId, $objId)
     {
         $this->db->manipulateF(
-            'DELETE FROM ' . self::DB_DEL_USERS . ' WHERE usr_id = %s',
-            ['integer'],
-            [$usrId]
+            'DELETE FROM ' . self::DB_DEL_USERS . ' WHERE usr_id = %s AND obj_id = %s',
+            ['integer', 'integer'],
+            [$usrId, $objId]
         );
     }
 
@@ -200,34 +201,33 @@ class ilCmiXapiDelModel
 
     public function setXapiObjAsDeleted(int $objId, int $typeId, string $actId): void
     {
-        $values = [
-            'obj_id' => ['integer', $objId],
-            'type_id' => ['integer', $typeId],
-            'activity_id' => ['string', $actId],
-            'added' => ['timestamp', date('Y-m-d H:i:s')]
-        ];
-        $this->db->insert(self::DB_DEL_OBJ, $values);
-
         if(!$this->dic->cron()->manager()->isJobActive('xapi_deletion_cron')) {
             $xapiDelete = new ilCmiXapiStatementsDeleteRequest($objId, $typeId, $actId, null, ilCmiXapiStatementsDeleteRequest::DELETE_SCOPE_ALL);
             $xapiDelete->delete();
+        } else {
+            $values = [
+                'obj_id' => ['integer', $objId],
+                'type_id' => ['integer', $typeId],
+                'activity_id' => ['string', $actId],
+                'added' => ['timestamp', date('Y-m-d H:i:s')]
+            ];
+            $this->db->insert(self::DB_DEL_OBJ, $values);
         }
     }
 
     public function setXapiObjAsDeletedForUser(int $objId, int $typeId, string $actId, int $usrId): void
     {
-        //        $values = [
-        //            'obj_id' => ['integer', $objId],
-        //            'type_id' => ['integer', $typeId],
-        //            'activity_id' => ['string', $actId],
-        //            'added' => ['timestamp', date('Y-m-d H:i:s')]
-        //        ];
-        //        $this->db->insert(self::DB_DEL_OBJ, $values);
-
-        //        if(!$this->dic->cron()->manager()->isJobActive('xapi_deletion_cron')) {
-        $xapiDelete = new ilCmiXapiStatementsDeleteRequest($objId, $typeId, $actId, $usrId, ilCmiXapiStatementsDeleteRequest::DELETE_SCOPE_ALL);
-        $xapiDelete->delete();
-        //        }
+        if(!$this->dic->cron()->manager()->isJobActive('xapi_deletion_cron')) {
+            $xapiDelete = new ilCmiXapiStatementsDeleteRequest($objId, $typeId, $actId, $usrId, ilCmiXapiStatementsDeleteRequest::DELETE_SCOPE_ALL);
+            $xapiDelete->delete();
+        } else {
+            $values = [
+                'usr_id' => ['integer', $usrId],
+                'obj_id' => ['integer', $objId],
+                'added' => ['timestamp', date('Y-m-d H:i:s')]
+            ];
+            $this->db->insert(self::DB_DEL_USERS, $values);
+        }
     }
